use koptinterface to reflow text in pdf/djvu readers
