+++ frontend/document/document.lua
@@ -254,5 +254,4 @@ end
 
 require "document/pdfdocument"
 require "document/djvudocument"
-require "document/koptdocument"
 require "document/credocument"
